+++ admin/pkg/selfupdate/deps.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package self_update
+package selfupdate
 
 import (
 	"context"
